# openapi.json
{
  "components": {
    "schemas": {
      "AppCreate": {
        "properties": {
          "brand_id": {
            "title": "Brand Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 255,
            "title": "Name",
            "type": "string"
          },
          "org_unit_id": {
            "title": "Org Unit Id",
            "type": "integer"
          },
          "slug": {
            "maxLength": 255,
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "brand_id",
          "org_unit_id"
        ],
        "title": "AppCreate",
        "type": "object"
      },
      "AppRead": {
        "properties": {
          "brand_id": {
            "title": "Brand Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_unit_id": {
            "title": "Org Unit Id",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "brand_id",
          "org_unit_id",
          "created_at",
          "updated_at"
        ],
        "title": "AppRead",
        "type": "object"
      },
      "AppUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "org_unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Unit Id"
          }
        },
        "title": "AppUpdate",
        "type": "object"
      },
...
      "Category": {
        "enum": [
          "aria",
          "color",
          "forms",
          "keyboard",
          "language",
          "name-role-value",
          "parsing",
          "semantics",
          "sensory-and-visual-cues",
          "structure",
          "tables",
          "text-alternatives",
          "time-and-media"
        ],
        "title": "Category",
        "type": "string"
      },
      "FindingType": {
        "enum": [
          "violation",
          "incomplete"
        ],
        "title": "FindingType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Impact": {
        "enum": [
          "critical",
          "serious",
          "moderate",
          "minor"
        ],
        "title": "Impact",
        "type": "string"
      },
      "NodeFindingDetail": {
        "properties": {
          "checks": {
            "additionalProperties": true,
            "title": "Checks",
            "type": "object"
          },
          "failure_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Summary"
          },
          "html": {
            "title": "Html",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "target": {
            "items": {
              "type": "string"
            },
            "title": "Target",
            "type": "array"
          }
        },
        "required": [
          "id",
          "html",
          "target",
          "impact",
          "failure_summary",
          "checks"
        ],
        "title": "NodeFindingDetail",
        "type": "object"
      },
...
      "Page_AppRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[AppRead]",
        "type": "object"
      },
...
      "Page_RuleFindingRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RuleFindingRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[RuleFindingRead]",
        "type": "object"
      },
...
      "Page_ScoreSnapshotRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScoreSnapshotRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ScoreSnapshotRead]",
        "type": "object"
      },
      "RuleFindingDetail": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "classifications": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Classifications",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "help": {
            "title": "Help",
            "type": "string"
          },
          "help_url": {
            "title": "Help Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "node_findings": {
            "items": {
              "$ref": "#/components/schemas/NodeFindingDetail"
            },
            "title": "Node Findings",
            "type": "array"
          },
          "page_result_id": {
            "title": "Page Result Id",
            "type": "integer"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "type": {
            "$ref": "#/components/schemas/FindingType"
          },
          "wcag_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Wcag Criteria",
            "type": "array"
          }
        },
        "required": [
          "id",
          "page_result_id",
          "rule_id",
          "type",
          "impact",
          "description",
          "help",
          "help_url",
          "category",
          "wcag_criteria",
          "classifications",
          "tags"
        ],
        "title": "RuleFindingDetail",
        "type": "object"
      },
      "RuleFindingRead": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "classifications": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Classifications",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "help": {
            "title": "Help",
            "type": "string"
          },
          "help_url": {
            "title": "Help Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "impact": {
            "$ref": "#/components/schemas/Impact"
          },
          "page_result_id": {
            "title": "Page Result Id",
            "type": "integer"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "type": {
            "$ref": "#/components/schemas/FindingType"
          },
          "wcag_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Wcag Criteria",
            "type": "array"
          }
        },
        "required": [
          "id",
          "page_result_id",
          "rule_id",
          "type",
          "impact",
          "description",
          "help",
          "help_url",
          "category",
          "wcag_criteria",
          "classifications",
          "tags"
        ],
        "title": "RuleFindingRead",
        "type": "object"
      },
...
      "ScoreSnapshotRead": {
        "properties": {
          "app_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id"
          },
          "avg_violations_per_page": {
            "title": "Avg Violations Per Page",
            "type": "number"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "org_unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Unit Id"
          },
          "pages_with_critical_violations": {
            "title": "Pages With Critical Violations",
            "type": "integer"
          },
          "pages_with_violations": {
            "title": "Pages With Violations",
            "type": "integer"
          },
          "pct_pages_with_critical_violations": {
            "title": "Pct Pages With Critical Violations",
            "type": "number"
          },
          "pct_pages_with_violations": {
            "title": "Pct Pages With Violations",
            "type": "number"
          },
          "scan_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Run Id"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "snapshot_at": {
            "format": "date-time",
            "title": "Snapshot At",
            "type": "string"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          },
          "total_violations": {
            "title": "Total Violations",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "score",
          "total_violations",
          "pages_with_violations",
          "pages_with_critical_violations",
          "total_pages",
          "avg_violations_per_page",
          "pct_pages_with_violations",
          "pct_pages_with_critical_violations",
          "id",
          "app_id",
          "scan_run_id",
          "org_unit_id",
          "brand_id",
          "snapshot_at",
          "created_at",
          "updated_at"
        ],
        "title": "ScoreSnapshotRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Accessibility Health API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/apps": {
      "get": {
        "operationId": "list_apps",
        "parameters": [
          {
            "in": "query",
            "name": "brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand Id"
            }
          },
          {
            "in": "query",
            "name": "org_unit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Org Unit Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AppRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Apps",
        "tags": [
          "apps"
        ]
      },
      "post": {
        "operationId": "create_app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create App",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/slug/{slug}": {
      "get": {
        "operationId": "get_app_by_slug",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get App By Slug",
        "tags": [
          "apps"
        ]
      }
    },
    "/api/v1/apps/{app_id}": {
      "delete": {
        "operationId": "delete_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete App",
        "tags": [
          "apps"
        ]
      },
      "get": {
        "operationId": "get_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get App",
        "tags": [
          "apps"
        ]
      },
      "patch": {
        "operationId": "update_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update App",
        "tags": [
          "apps"
        ]
      }
    },
...
    "/api/v1/apps/{app_id}/scores": {
      "get": {
        "operationId": "list_app_scores",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ScoreSnapshotRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List App Scores",
        "tags": [
          "apps"
        ]
      }
    },
...
    "/api/v1/scan-runs/{scan_run_id}/findings": {
      "get": {
        "operationId": "list_findings",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/FindingType"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "impact",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Impact"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Impact"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "wcag_criterion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wcag Criterion"
            }
          },
          {
            "in": "query",
            "name": "classification",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "wcag2a",
                      "wcag2aa",
                      "wcag21a",
                      "wcag21aa",
                      "wcag22a",
                      "wcag22aa",
                      "best-practice"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Classification"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RuleFindingRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Findings",
        "tags": [
          "findings"
        ]
      }
    },
    "/api/v1/scan-runs/{scan_run_id}/findings/{finding_id}": {
      "get": {
        "operationId": "get_finding",
        "parameters": [
          {
            "in": "path",
            "name": "scan_run_id",
            "required": true,
            "schema": {
              "title": "Scan Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleFindingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Finding",
        "tags": [
          "findings"
        ]
      }
    },
...
  }
}